remove tests, fix existing test
