+++ hapi-fhir-docs/src/main/resources/ca/uhn/hapi/fhir/changelog/6_8_0/5000-revincludes-not-tracked-in-sql-tracing.yaml
@@ -1,0 +1,4 @@
+---
+type: fix
+issue: 5000
+title: "Previously, the JPA_PERFTRACE_RAW_SQL was not firing for raw sql that occurred during a `revinclude` or `include`. This has been corrected."
