+++ hapi-fhir-jpaserver-test-r4/src/test/java/ca/uhn/fhir/jpa/provider/r4/ResourceProviderR4Test.java
@@ -277,87 +277,6 @@ public class ResourceProviderR4Test extends BaseResourceProviderR4Test {
 		myStorageSettings.setSearchPreFetchThresholds(new JpaStorageSettings().getSearchPreFetchThresholds());
 	}
 
-
-
-	class SqlCapturingInterceptor {
-		SqlQueryList queryList = null;
-		@Hook(Pointcut.JPA_PERFTRACE_RAW_SQL)
-		public void trackRequest(RequestDetails theRequestDetails, SqlQueryList theQueryList) {
-			if (queryList == null) {
-				queryList = theQueryList;
-			} else {
-				queryList.addAll(theQueryList);
-			}
-		}
-
-		public SqlQueryList getQueryList() {
-			return queryList;
-		}
-	}
-	@Test
-	public void testRawSqlCapturingGetsIncludesAndRevIncludes() {
-
-		SqlCapturingInterceptor interceptor = new SqlCapturingInterceptor();
-		try {
-			Patient p = new Patient();
-			IIdType unqualified = myClient.create().resource(p).execute().getId().toUnqualified();
-			Observation o = new Observation();
-			o.setSubject(new Reference(unqualified));
-			myClient.create().resource(o).execute();
-
-			myInterceptorRegistry.registerInterceptor(interceptor);
-			Bundle execute = myClient.search().byUrl("Patient?_revinclude=Observation:subject").returnBundle(Bundle.class).execute();
-			assertThat(execute.getEntry(), hasSize(2));
-			SqlQueryList queryList = interceptor.getQueryList();
-			//Basic query, then the revinclude
-			assertThat(queryList, hasSize(2));
-		} finally {
-			myInterceptorRegistry.unregisterInterceptor(interceptor);
-		}
-	}
-
-	@Test
-	public void testRawSqlCapturingGetsIteratedRevIncludes() {
-
-		SqlCapturingInterceptor interceptor = new SqlCapturingInterceptor();
-		try {
-			Patient p = new Patient();
-			IIdType patientId = myClient.create().resource(p).execute().getId().toUnqualified();
-
-			Condition condition = new Condition();
-			condition.setSubject(new Reference(patientId));
-			myClient.create().resource(condition).execute();
-
-			Group g = new Group();
-			g.addMember().setEntity(new Reference(patientId));
-			IIdType groupId = myClient.create().resource(g).execute().getId().toUnqualified();
-
-			CareTeam careTeam = new CareTeam();
-			careTeam.setSubject(new Reference(groupId));
-			myClient.create().resource(careTeam).execute();
-
-			DetectedIssue detectedIssue = new DetectedIssue();
-			detectedIssue.setPatient(new Reference(patientId));
-			myClient.create().resource(detectedIssue).execute();
-
-			myInterceptorRegistry.registerInterceptor(interceptor);
-
-			Bundle execute = myClient.search()
-				.byUrl("Patient?_id="+ patientId.getIdPart() +
-					"&_revinclude=Condition:patient" +
-					"&_revinclude:iterate=CareTeam:subject" +
-					"&_revinclude=Group:member:Patient" +
-					"&_revinclude=DetectedIssue:patient"
-				).returnBundle(Bundle.class).execute();
-			assertThat(execute.getEntry(), hasSize(2));
-			SqlQueryList queryList = interceptor.getQueryList();
-			//Basic query, then the revinclude
-			assertThat(queryList, hasSize(5));
-		} finally {
-			myInterceptorRegistry.unregisterInterceptor(interceptor);
-		}
-	}
-
 	@Test
 	public void testParameterWithNoValueThrowsError_InvalidChainOnCustomSearch() throws IOException {
 		SearchParameter searchParameter = new SearchParameter();
